Change dependency scopes for instrumentation libraries, to achieve optionality
* Since the instrumentation-runner library is automatically applied
  by the Gradle plugin, it cannot put any JUnit 5 dependencies
  on the runtime classpath itself. That would result in a premature
  minSdkVersion bump for all target apps, and has to be avoided.

* Instead, the instrumentation-runner only compiles against them,
  and the remaining artifacts provide the APIs at runtime for only
  the flavors that utilize JUnit 5 instrumentation tests.

* The sample project has been enhanced to demonstrate a potential
  integration of JUnit 5 instrumentation tests, using a productFlavor.
