+++ sample/build.gradle
@@ -24,12 +24,24 @@ android {
     versionName "1.0"
 
     testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+    testInstrumentationRunnerArgument "runnerBuilder", "de.mannodermaus.junit5.AndroidJUnit5Builder"
   }
 
-  sourceSets {
-    androidTest.java.srcDirs += "src/androidTest/kotlin"
-    main.java.srcDirs += "src/main/kotlin"
-    test.java.srcDirs += "src/test/kotlin"
+  flavorDimensions "kind"
+  productFlavors {
+    experimental {
+      dimension "kind"
+      minSdkVersion 26
+    }
+
+    normal {
+      dimension "kind"
+    }
+  }
+
+  // Add Kotlin source directory to all source sets
+  sourceSets.each {
+    it.java.srcDirs += "src/$it.name/kotlin"
   }
 
   compileOptions {
@@ -55,4 +67,12 @@ dependencies {
   testImplementation junit5.unitTests()
   testImplementation junit5.parameterized()
   testCompileOnly junit5.unitTestsRuntime()
+
+  androidTestImplementation "junit:junit:$JUNIT4_VERSION"
+  androidTestImplementation "com.android.support.test:runner:$SUPPORT_TEST_LIBRARY_VERSION"
+
+  androidTestRuntimeOnly project(":instrumentation-runner")
+  androidTestExperimentalImplementation junit5.unitTests()
+  androidTestExperimentalImplementation junit5.parameterized()
+  androidTestExperimentalImplementation project(":instrumentation")
 }
