+++ sample/src/androidTest/kotlin/de/mannodermaus/sample/TestRunningOnJUnit4.kt
@@ -1,0 +1,11 @@
+package de.mannodermaus.sample
+
+import org.junit.Assert
+import org.junit.Test
+
+class TestRunningOnJUnit4 {
+  @Test
+  fun junit4() {
+    Assert.assertEquals(4, 2 + 2)
+  }
+}
