+++ sample/src/main/AndroidManifest.xml
@@ -1,2 +1,8 @@
 <!-- Nothing to see here, check out the "test" source set instead! -->
-<manifest package="de.mannodermaus.junit5.sample"/>
+<manifest xmlns:android="http://schemas.android.com/apk/res/android"
+    package="de.mannodermaus.junit5.sample">
+
+  <application>
+    <activity android:name=".ActivityOne"/>
+  </application>
+</manifest>
