+++ sample/src/main/kotlin/de/mannodermaus/junit5/sample/ActivityOne.kt
@@ -1,0 +1,7 @@
+package de.mannodermaus.junit5.sample
+
+import android.app.Activity
+
+class ActivityOne : Activity() {
+
+}
